teleport the legacy SMS views over to the dedicated application
Ideally we'll be able to switch off legacy views at some point. We
should start with a view to making that process easy by restricting them
to their own application.
