+++ legacysms/urls.py
@@ -10,5 +10,6 @@ from . import views
 app_name = "legacy-sms"
 
 urlpatterns = [
-    path('example', views.example, name='example'),
+    path('embed/<int:media_id>/', views.embed, name='embed'),
+    path('rss/media/<int:media_id>/', views.rss_media, name='rss_media'),
 ]
