+++ smsjwplatform/urls.py
@@ -15,12 +15,5 @@ Default URL patterns for the :py:mod:`smsjwplatform` application are provided by
 
 """
 
-from django.urls import path
-
-from . import views
-
 app_name = 'smsjwplatform'
-urlpatterns = [
-    path('embed/<int:media_id>/', views.embed, name='embed'),
-    path('rss/media/<int:media_id>/', views.rss_media, name='rss_media'),
-]
+urlpatterns = []
